+++ src/clipboard.ts
@@ -11,3 +11,28 @@ export const writeViaNavigator = ({
 
   return navigator.clipboard.write([new ClipboardItem(items)]);
 };
+
+export const writeViaCommand = ({
+  text,
+  html,
+}: {
+  text?: string | null;
+  html?: string | null;
+}): Promise<void> => {
+  const handleCopy = (event: ClipboardEvent) => {
+    event.preventDefault();
+    if (text) event.clipboardData?.setData("text/plain", text);
+    if (html) event.clipboardData?.setData("text/html", html);
+  };
+  document.addEventListener("copy", handleCopy);
+  return new Promise<void>((resolve, reject) => {
+    try {
+      document.execCommand("copy");
+      resolve();
+    } catch (error) {
+      reject(error);
+    } finally {
+      document.removeEventListener("copy", handleCopy);
+    }
+  });
+};