+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add a direct keyboard shortcut to copy a Markdown link to the current page.
 - Copy the text/html version of a link in addition when copying a Markdown link.
 - Fix Markdown escaping.